Fix webpush not success response
